requirements.txt
apache-airflow==2.10.5
finnhub-python==2.4.27
psycopg2-binary==2.9.11
python-dotenv==1.2.2
websocket-client==1.9.0
